Move query client out of experimental
